Move camera & viewport into a separate directory
