+++ packages/flame/lib/game.dart
@@ -1,10 +1,10 @@
 export 'src/extensions/vector2.dart';
 export 'src/fps_counter.dart';
 export 'src/game/base_game.dart';
-export 'src/game/camera.dart';
+export 'src/game/camera/camera.dart';
+export 'src/game/camera/viewport.dart';
 export 'src/game/game.dart';
 export 'src/game/game_widget/game_widget.dart';
 export 'src/game/projector.dart';
 export 'src/game/transform2d.dart';
-export 'src/game/viewport.dart';
 export 'src/text.dart';
