+++ packages/flame/lib/src/game/camera/viewport.dart
@@ -3,9 +3,9 @@ import 'dart:ui';
 
 import 'package:flutter/painting.dart';
 
-import '../../extensions.dart';
-import '../../game.dart';
-import 'projector.dart';
+import '../../../extensions.dart';
+import '../../../game.dart';
+import '../projector.dart';
 
 /// A viewport is a class that potentially translates and resizes the screen.
 /// The reason you might want to have a viewport is to make sure you handle any
